Abstracted the userDashboard component to a separate folder in the components directory called Dashboards. Will make the dashboard more dynamic so the same structure can be applied to both the admin dashboard and the user dashboard.
