+++ client/src/containers/Admin/AdminDashboard.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import Dashboard from '../../components/Dashboards/Dashboard';
+
+const AdminDashboard = () => {
+    return (
+        <div>
+            <Dashboard />
+        </div>
+    );
+};
+
+export default AdminDashboard;
