+++ client/src/containers/UserDashboard/UserDashboard.jsx
@@ -1,61 +1,11 @@
-import React, { useEffect, useState } from "react";
-import axios from 'axios';
-
-const styles = {
-  thead: {
-    backgroundColor: "#bdbdbd"
-  }
-}
-
+import React from 'react';
+import Dashboard from '../../components/Dashboards/Dashboard';
 
 const UserDashboard = () => {
-const [subscription_thumbnail, setSubcription_thumbnail] = useState("");
-const [subscription_price, setSubscription_price] = useState("");
-
-axios.get('/api/subscriptions', (req, res) => {
-
-}).then((response) => {
-  console.log(response.data);
-}).catch((err) => {
-console.log(err)
-});
-
-
   return (
-    <>
-    <br/>
-    <br/>
     <div>
-    <div className="container">
-    <div className="row">
-    <div className="col s12">
-    <table>
-        <thead style={styles.thead}>
-          <tr>
-            <th>Subscription Name</th>
-            <th>Image</th>
-            <th>Current Amount</th>
-            <th>Subscription Price Change</th>
-            <th>Update Subscription</th>
-            <th>Delete Subscription</th>
-          </tr>
-        </thead>
-        <tbody>
-          <tr>
-            <td>Name</td>
-            <td>Image</td>
-            <td>$13.99</td>
-            <td>$1.00</td>
-            <td><a href="/" className="waves-effect waves-light btn-small"><i className="material-icons left">update</i>UPDATE</a></td>
-            <td><a href="/" className="waves-effect waves-light btn-small"><i className="material-icons left">delete</i>DELETE</a></td>
-          </tr>
-        </tbody>
-      </table>
+      <Dashboard />
     </div>
-    </div>
-    </div>
-    </div>
-    </>
   );
 };
